update exports, remove core, fix component, new storybook story
