+++ packages/manager/src/components/AccessPanel/UserSSHKeyPanel.tsx
@@ -1,7 +1,7 @@
 import { Theme } from '@mui/material/styles';
 import * as React from 'react';
 import { Button } from 'src/components/Button/Button';
-import CheckBox from 'src/components/CheckBox';
+import { Checkbox } from 'src/components/Checkbox';
 import { TableBody } from 'src/components/TableBody';
 import { TableHead } from 'src/components/TableHead';
 import { Typography } from 'src/components/Typography';
@@ -130,7 +130,7 @@ const UserSSHKeyPanel = (props: Props) => {
       return (
         <TableRow>
           <TableCell className={classes.cellCheckbox}>
-            <CheckBox
+            <Checkbox
               disabled={disabled}
               checked={authorizedUsers.includes(profile.username)}
               onChange={() => onToggle(profile.username)}
@@ -162,7 +162,7 @@ const UserSSHKeyPanel = (props: Props) => {
     return users?.data.map((user) => (
       <TableRow key={user.username}>
         <TableCell className={classes.cellCheckbox}>
-          <CheckBox
+          <Checkbox
             disabled={disabled || user.ssh_keys.length === 0}
             checked={authorizedUsers.includes(user.username)}
             onChange={() => onToggle(user.username)}
