--- packages/manager/src/components/CheckBox/index.ts
@@ -1,2 +1,0 @@
-import CheckBox from './CheckBox';
-export default CheckBox;
