--- packages/manager/src/components/core/Card.ts
@@ -1,6 +1,0 @@
-import Card, { CardProps as _CardProps } from '@mui/material/Card';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface CardProps extends _CardProps {}
-
-export default Card;
