--- packages/manager/src/components/core/CardActions.ts
@@ -1,8 +1,0 @@
-import CardActions, {
-  CardActionsProps as _CardActionsProps,
-} from '@mui/material/CardActions';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface CardActionsProps extends _CardActionsProps {}
-
-export default CardActions;
