--- packages/manager/src/components/core/CardContent.ts
@@ -1,8 +1,0 @@
-import CardContent, {
-  CardContentProps as _CardContentProps,
-} from '@mui/material/CardContent';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface CardContentProps extends _CardContentProps {}
-
-export default CardContent;
