--- packages/manager/src/components/core/CardHeader.ts
@@ -1,8 +1,0 @@
-import CardHeader, {
-  CardHeaderProps as _CardHeaderProps,
-} from '@mui/material/CardHeader';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface CardHeaderProps extends _CardHeaderProps {}
-
-export default CardHeader;
