--- packages/manager/src/components/core/Checkbox.ts
@@ -1,8 +1,0 @@
-import Checkbox, {
-  CheckboxProps as _CheckboxProps,
-} from '@mui/material/Checkbox';
-
-/* tslint:disable-next-line:no-empty-interface */
-export interface CheckboxProps extends _CheckboxProps {}
-
-export default Checkbox;
