+++ packages/manager/src/features/Profile/OAuthClients/CreateOAuthClientDrawer.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import ActionsPanel from 'src/components/ActionsPanel';
 import { Button } from 'src/components/Button/Button';
-import CheckBox from 'src/components/CheckBox';
+import { Checkbox } from 'src/components/Checkbox';
 import FormControl from 'src/components/core/FormControl';
 import FormControlLabel from 'src/components/core/FormControlLabel';
 import Drawer from 'src/components/Drawer';
@@ -67,7 +67,7 @@ export const CreateOAuthClientDrawer = ({
           <FormControlLabel
             label="Public"
             control={
-              <CheckBox
+              <Checkbox
                 name="public"
                 onChange={formik.handleChange}
                 checked={formik.values.public}
