+++ packages/manager/src/features/Profile/OAuthClients/EditOAuthClientDrawer.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import ActionsPanel from 'src/components/ActionsPanel';
 import { Button } from 'src/components/Button/Button';
-import CheckBox from 'src/components/CheckBox';
+import { Checkbox } from 'src/components/Checkbox';
 import FormControl from 'src/components/core/FormControl';
 import FormControlLabel from 'src/components/core/FormControlLabel';
 import Drawer from 'src/components/Drawer';
@@ -71,7 +71,7 @@ export const EditOAuthClientDrawer = ({ open, onClose, client }: Props) => {
           <FormControlLabel
             label="Public"
             control={
-              <CheckBox name="public" checked={client?.public} disabled />
+              <Checkbox name="public" checked={client?.public} disabled />
             }
           />
         </FormControl>
